Expose LedDevice getLedCount and align signedness

+++ libsrc/leddevice/dev_hid/LedDeviceMultiLightpack.cpp
@@ -101,7 +101,7 @@ int LedDeviceMultiLightpack::write(const std::vector<ColorRgb> &ledValues)
 
 	for (LedDeviceLightpack * device : _lightpacks)
 	{
-		int count = qMin(device->getLedCount(), size);
+		int count = qMin(static_cast<int>( device->getLedCount()), size);
 
 		if (count > 0)
 		{
